enforce simple repository structure
Supporting paekli-rs as a dedicated repository as well as in a subdirectory
of another repository is proving too complicated.

Either the shell commands have to be more complicated, which is confusing and
distracting to users. Or the instructions constantly have to tell people to
adjust paths in some commands, which is distracting, annoying and error prone.